requirements-dev.txt
conan
cmake
flake8
lxml
mypy
macholib ; sys_platform == 'darwin'
pybind11
pytest
pytest-runner
sphinx
tox
